+++ Filters/PlatformResourceFilter.cs
@@ -1,10 +1,12 @@
 using System;
+using System.Collections.Generic;
 using System.IO.Pipelines;
 using System.Linq;
 using System.Text;
 using System.Text.Json;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Filters;
+using Microsoft.Extensions.Primitives;
 using RestSharp.Serialization.Json;
 using Rumble.Platform.Common.Utilities;
 using Rumble.Platform.Common.Web;
@@ -33,13 +35,25 @@ namespace Rumble.Platform.Common.Filters
 			}
 			try
 			{
-				if (context.HttpContext.Request.Method == "GET")
-					return;
-				context.HttpContext.Request.BodyReader.TryRead(out ReadResult result);
-				string json = Encoding.UTF8.GetString(result.Buffer.FirstSpan);
-				context.HttpContext.Request.BodyReader.AdvanceTo(result.Buffer.End);
-				context.HttpContext.Request.BodyReader.Complete();
-				context.HttpContext.Items[KEY_BODY] = JsonDocument.Parse(json, JsonHelper.DocumentOptions);
+				GenericData query = new GenericData();
+				GenericData body = null;
+				
+				foreach (KeyValuePair<string, StringValues> pair in context.HttpContext.Request.Query)
+					query[pair.Key] = pair.Value.ToString();
+				if (context.HttpContext.Request.Method != "GET")
+				{
+					context.HttpContext.Request.BodyReader.TryRead(out ReadResult result);
+					string json = Encoding.UTF8.GetString(result.Buffer.FirstSpan);
+					context.HttpContext.Request.BodyReader.AdvanceTo(result.Buffer.End);
+					context.HttpContext.Request.BodyReader.Complete();
+					body = json;
+				}
+
+				body?.Combine(query); // If both the body and query have the same key, the values in the body have priority.
+				body ??= query;
+				
+				context.HttpContext.Items[KEY_BODY] = body;
+				// context.HttpContext.Items[KEY_BODY] = JsonDocument.Parse(json, JsonHelper.DocumentOptions);
 			}
 			catch (Exception e)
 			{